lab6/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-jose[cryptography]==3.4.0
bcrypt==5.0.0
python-multipart==0.0.20
